+++ src/main.cpp
@@ -1,11 +1,7 @@
 #include "mainwindow.h"
 #include <QApplication>
 #include <QtGui>
-#include <QTextEdit>
-#include <QSplitter>
-#include <QVBoxLayout>
-#include <QListView>
-#include <QSystemTrayIcon>
+#include <QTranslator>
 
 
 int main(int argc, char *argv[])
@@ -16,6 +12,14 @@ int main(int argc, char *argv[])
 #endif
 
     QApplication a(argc, argv);
+
+    QString locale = QLocale::system().name().section('_', 0, 0);
+
+    QTranslator translator;
+    // TODO(pbek): this has to be adapted for build systems
+    translator.load("../src/languages/QOwnNotes_" + locale);
+
+    a.installTranslator(&translator);
 
     MainWindow w;
     w.show();