+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # QOwnNotes Changelog
 
 ## 0.87
-- start of app translation support implementation
+- start of app internationalization implementation
   (for [Issue #40](https://github.com/pbek/QOwnNotes/issues/40))
 - fixed a problem with the crypto key expiration of encrypted messages where one
   had to enter the password every minute while editing the note
   (for [Issue #42](https://github.com/pbek/QOwnNotes/issues/42))
-    - now the crypto key will expire after 10min without editing the note  
+    - now the crypto key will expire after 10min without editing the note
+- did some fixes and layout changes in the settings dialog
 - added basic app metrics
     - anonymous usage data helps to decide what parts of QOwnNotes to 
     improve next and to find and fix bugs